bookanalyzer/requirements.txt
annotated-types==0.7.0
asgiref==3.9.1
beautifulsoup4==4.13.5
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
colorama==0.4.6
Django==5.2.6
djangorestframework==3.16.1
faiss-cpu==1.12.0
filelock==3.19.1
fsspec==2025.9.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.182.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.75.0
grpcio-status==1.71.2
httplib2==0.31.0
huggingface-hub==0.35.0
idna==3.10
Jinja2==3.1.6
joblib==1.5.2
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pillow==11.3.0
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.9
pydantic_core==2.33.2
pyparsing==3.2.4
PyPDF2==3.0.1
pytesseract==0.3.13
python-decouple==3.8
PyYAML==6.0.2
regex==2025.9.1
requests==2.32.5
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
sentence-transformers==5.1.0
soupsieve==2.8
sqlparse==0.5.3
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.0
torch==2.8.0
tqdm==4.67.1
transformers==4.56.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
